requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openai==1.58.1
python-dotenv==1.0.1
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
pydantic
pydantic-core
azure-identity==1.19.0
azure-search-documents==11.6.0b4
numpy==2.1.3
pandas==2.2.3
python-multipart==0.0.17
aiofiles==24.1.0
tf-keras
